+++ druid/src/widget/container.rs
@@ -106,7 +106,10 @@ impl<T: Data> Widget<T> for Container<T> {
         self.inner.lifecycle(ctx, event, data, env)
     }
 
-    fn update(&mut self, ctx: &mut UpdateCtx, _old_data: &T, data: &T, env: &Env) {
+    fn update(&mut self, ctx: &mut UpdateCtx, old_data: &T, data: &T, env: &Env) {
+        if let Some(BackgroundBrush::Painter(p)) = self.background.as_mut() {
+            p.update(ctx, old_data, data, env);
+        }
         self.inner.update(ctx, data, env);
     }