+++ CHANGELOG.md
@@ -111,6 +111,7 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - Focus request handling is now predictable with the last request overriding earlier ones. ([#948] by [@xStrom])
 - Wheel events now properly update hot state. ([#951] by [@xStrom])
 - X11: Support mouse scrolling. ([#961] by [@jneem])
+- Painter widget not repainting on data change in Container ([#991] by [@cmyr])
 
 ### Visual
 
@@ -231,6 +232,7 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 [#970]: https://github.com/xi-editor/druid/pull/970
 [#980]: https://github.com/xi-editor/druid/pull/980
 [#982]: https://github.com/xi-editor/druid/pull/982
+[#991]: https://github.com/xi-editor/druid/pull/991
 
 ## [0.5.0] - 2020-04-01